Backend/requirements.txt
fastapi
python-multipart
uvicorn
replicate
pydantic
openai
numpy
Pillow
pytesseract
validators
requests
boto3